zip-create: remove `images` in zip path
The tool must not alter file paths between the local directory and what
will be served by the Ghost instance.

On the "Copy the directory" task, an `/images` is added to the
destination temporary path. This addition changes the path of files that
is preserved by the zipping process.
This results in an incorrect path on the Ghost instance once uploaded,
as the import process preserves paths.