zip-create: await zip writing before ending tasks
The `createZip` function, called for each chunk, is asynchronous. While
`createZip` calls are correctly awaited within `Promise.all()`, the
internal call to `fsUtils.zip.write` is not awaited. This function is
also asynchronous:
https://github.com/TryGhost/migrate/blob/7686d2d3be174fde8ebbc78699aa7b3e7b89d7e4/packages/mg-fs-utils/lib/zip.js#L57

This omission caused the Zipping chunks task to exit prematurely,
without waiting the end of zip file creation.
With this fix the comment is not needed and the cleanup task can be
re-activated.